Floating IP does not work with VLAN tenant networks
When trying to use the Floating IP address of an instance
in a OVN-DVR setup while using VLAN tenant network type
the traffic does not pass. This is because of limitations
with VLAN tenant network types, which doesn't support
distributed VLAN routing.

The problem is the external_mac should not be set in the
NAT table for the Floating IP which belongs to the logical
port of a VLAN tenant logical switch. Otherwise, the compute
node which is binding the logical port will send the GARP
where as the routing for VLAN tenant networks is centralized.

Change-Id: I406204fc751387c02517286c87fe45533d70b93d
Closes-bug: #1828891
(cherry picked from commit c5aef51edc9843db605303ec8bd8610b6c55e9c2)