Expand the tempest_test_regex variable
The tempest_test_regex variable from the networking-ovn-jobs.yaml is
just a long string with no line breaks which makes it really hard to
read to know what is enabled or disabled in the tempest tests.

This patch expands that variable to make it more readable and easy to
expand/contract if needed.

TrivialFix

Change-Id: I600c4d22db71a08b1576bb583620b525aa529264
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>