functional: Wait for PB event before binding the port
On busy boxes there is a chance ovn-northd doesn't create a port binding
entry quick enough in the Southbound DB when a new logical switch port
is created. It can happen that test attempts to bind a port that doesn't
exist in the southbound DB yet, because it's being processed by
ovn-northd.

This patch adds a one time event and the test waits 5 seconds for the
port binding to be created. Only after that it calls lsp_bind().

The patch also makes the lsp_bind fails in case of failure inside of
ovsdbapp, previously, the exception was swallowed and it made debugging
a lot harder.

It was required to bump ovsdbapp version from 0.10.0 to 0.14.0 because
that's the version that implements WaitEvent class.

UPDATE: Rocky has ovsdbapp version of 0.12.3 - this patch bumps the
        version to 0.12.4 that is about to be released and contains the
        change this patch requires.

Closes bug: #1830221

Depends-On: https://review.opendev.org/#/c/661797

Change-Id: Ieb48056ef728b12ced01f2829dc2d57b9bb6b4dc
(cherry picked from commit e9c4a51e8179245eb308d90e1b35baa914227e64)
(cherry picked from commit 57160e81b50b93dbdf64b92c0b4aa5dead7effe2)