functional: Wait for PB event before binding the port
On busy boxes there is a chance ovn-northd doesn't create a port binding
entry quick enough in the Southbound DB when a new logical switch port
is created. It can happen that test attempts to bind a port that doesn't
exist in the southbound DB yet, because it's being processed by
ovn-northd.

This patch adds a one time event and the test waits 5 seconds for the
port binding to be created. Only after that it calls lsp_bind().

The patch also makes the lsp_bind fails in case of failure inside of
ovsdbapp, previously, the exception was swallowed and it made debugging
a lot harder.

It was required to bump ovsdbapp version from 0.10.0 to 0.14.0 because
that's the version that implements WaitEvent class.

Closes bug: #1830221

Change-Id: Ieb48056ef728b12ced01f2829dc2d57b9bb6b4dc