functional: Import event module from ovsdbapp
Originally the code imported the whole module including tests for
ovsdbapp. After [1] is merged, we no longer need those tests and can
import only event module.

This patch attempts to load the new module first. Because the module was
introduced in [1], in case of older ovsdbapp version, the import will
fail. In such case it falls back to original location.

[1] https://review.opendev.org/#/c/661632/

Change-Id: I8ad44a236062d23df520646396e459208404a061