Omit Octavia code from coverage report
There is a patch in progress [0] to increase the test
coverage in the Octavia code. Until then, omit it from
the report so the coverage test will pass, so we can
make it voting.
For now we have test coverage at level of 66%.
As a good start set 65% to start voting and then
successfully increase it.

[0] https://review.opendev.org/#/c/601948/

Co-Authored-By: Maciej Józefczyk <mjozefcz@redhat.com>
Change-Id: I0d63fc127da9ab37bb217678a1c47961c17d3eda