Hash Ring: Make workers health check more reliable
When under pressure the maintenance thread may not honor the time interval
that it should run its tasks. For something more sensitive like the task
that health checks the nodes from the Hash Ring this can be problematic
because it may cause it to appear that the workers have died and the
ring will get rebalanced; we do not want to have a long interval for it
because in case of a real failure we do want the ring to rebalance fast.

This patch implements a new approach to enhance the health check of the
Hash Ring workers by allowing the worker to "touch" the ring to say that
its alive.

The patch also includes unittests for the OvnIdlDistributedLock class
that was missing before.

Closes-Bug: #1834498
Change-Id: If06d8580e8e2637b19ff5d76e16f635a5c1d328e
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>