Functional tests: Fix setup error due to missing function
The "fixup_uca" was recently renamed in DevStack to "fixup_ubuntu"
as part of https://review.opendev.org/#/c/619562 and that broke our
functional gate setup. The functional itself does not seeing to be
needed anymore (functional tests passes without it) so this patch is
getting rid of it to avoid another renaming in the future.

Change-Id: I15419c029cd9e60a1a6164f97fa893611994b8f3
Closes-Bug: #1835375
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>