Hash Ring: Add support for groups
This patch adds support for groups in the Hash Ring mechanism.

Prior to this patch there was an assumption that all members of the ring
were able to handle the same events but, once we start expanding the use
of the hash ring on different parts of the code such as the Octavia
driver, that assumption is not true anymore because they have their own
events.

With the group mechanism, multiple groups can be created (resulting in a
separated hash ring for each group) that way we can group workers based
on the events they are able to handle.

This patch includes a new migration script that adds a new column called
"group_name" to the "ovn_hash_ring" table.

Change-Id: Ifc489afb764cddeec58597399d7d641c0d7ff279
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>