Always add NAT rule to a LR when updating a FIP
Before this patch, the NAT rule was attempted to be updated when a
FIP was reassigned to another port. However, this was a noop
since the NAT rule didn't exist anymore causing the reassigning to
be ineffective and to fail silently.

This patch is always adding the NAT rule to the Logical Router
no matter if the Neutron FIP is being added or updated.

The bug that this patch addresses was being hit in the gate with
around a 30% ratio because the maintenance task was fixing it while
tempest was still rying to SSH into it.

Conflicts:
	networking_ovn/common/ovn_client.py

Change-Id: Icebf4a82f64989112c3ca810b4358de490108c2d
Closes-Bug: #1835029
Closes-Bug: #1833820
Co-Authored-By: Jakub Libosvar <jlibosva@redhat.com>
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit d662f444280f5dc0a304a0b4de4767c91304c747)
(cherry picked from commit 13689522d6b97b4d2246e5de6af0447b39ec50ed)