OVN to emit ICMP "fragmentation needed" packets
A new configuration option called "ovn_emit_need_to_frag" has been added
and defaults to False. Reason being, it requires a specific kernel
version for it to work (upstream >= 5.2). It's also possible to find out
if OVN supports that operation by running the following command as root:

$ ovs-appctl -t ovs-vswitchd dpif/show-dp-features br-int
(check for "Check pkt length action: Yes").

WARNING: Enabling this option without kernel support will result in a
real performance degradation.

At earlier versions of this patch we thought about having networking-ovn
to discover if this option is supported or not (by issuing that
ovs-appctl command above) but, since networking-ovn doesn't run in the
network nodes we thought that an upper level tool (such as puppet-ovn or
TripleO) could do a better job at checking for it and configuring things
accordingly.

Closes-Bug: #1838405
Change-Id: I089f95b40803a6cd5e01990acacd599ced3bbd91
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>