DevStack: Use OVN from ovn-org/ovn (the new OVN repository)
This patch is modifying the DevStack scripts to use the new OVN
repository for our jobs targeting the master branch (or >= 2.12, when
it's released).

If the OVN_BRANCH is set to stable branches the code will fallback to
using the "compile_ovs" function like before (since those branches does
not exist in the new OVN repository, only master and >= 2.12 when
released).

Change-Id: I8b99a3481b4460ae32f2258dcf4190db9a8fed89
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>