Handle events in separate IDL instance
While setting the event handler lock in OvnNbIdlForLb
we found that on each Octavia API call exception
was raised that the IDL has not aquired OVNDB lock [0].

It is because when we set the event lock name it is
required to be acquired on each transaction, even on
the one passed by Octavia API, which shouldn't be
blocking one [1].

The second issue was cached NB DB connection within the same
process [2]. We overcome this by using other idl
implementation that utilize modified
networking_ovn.ovsdb.impl_idl_ovn.Backend.

So to address this issue I initialize second IDL that
watches only for events and has the lock name set. First
instance of OvnNbIdlForLb is used only for handling
Octavia API requests. Second has the lock name set and
is cached by OvnProviderHelper class attribute, that makes
the event mechanism work.

While long-living agent driver would be released
the second OvnNbIdlForLb object could be used in OVN agent
driver process in order to handle events there and then
we could stop using second IDL in the OvnProviderHelper.

[0] https://github.com/openstack/ovsdbapp/blob/master/ovsdbapp/backend/ovs_idl/transaction.py#L102
[1] https://github.com/openvswitch/ovs/blob/master/python/ovs/db/idl.py#L445
[2] https://github.com/openstack/ovsdbapp/blob/68d462d355a9c9f7e6b12f08f4dfb0b0b3d443b2/ovsdbapp/backend/ovs_idl/__init__.py#L51

Change-Id: I6534ac7777cf2a1216c65110704c675d5160b051
Story: 2006196