Add missing unittests to OVN provider driver
This change bumps up test coverage.
During writing those tests I found also a few issues:

1) In method pool_update the operating_status shouldn't
be set with veryfing if the pool has members or not.
The only way of setting operating_status to OFFLINE is
to administratively disable it or via member healt monitor,
which we don't yet support. Dropping this wrong code.
https://docs.openstack.org/api-ref/load-balancer/v2/?expanded=create-pool-detail#create-pool

2) _check_monitor_options() logic was wrong - fixed it

3) In pool_delete() in some coditions the external_ids
could be used, but not initlized. I moved it to be on the top
of function.

Change-Id: Ib60ba2b7a958e3d55699988d9d4f2b848e5e1b5a
(cherry picked from commit 77d03f6f17081e0f8c15cab0444261bc4e517055)