Fix evacuation when host dies uncleanly
If a host crashes and ovn-controller doesn't clean up the
Port_Binding chassis column, the Logical_Switch_Port is never set
to DOWN, so we do not detect the up->down transition and update the
port status with the driver. This patch watches for Port_Binding
chassis column changes and if the port is up goes ahead and updates
the driver.

Change-Id: I8a5e6ad2e98b79a140977ce003609ed5b21e3499
Closes-Bug: #1840876