functional: Fix test_router tests LSP binding
On busy machines when running funcional tests, it may take time to
process new logical port and create port bindings entry in the
southbound database.

Because the tests need to wait for gateway router port that is created
along with router object, we don't know the name of the logical port
beforehand. The patch creates a new wait event that matches on ports
starting with "cr-lrp-" string and caches all such ports. Later, the
wait() method implements parameter for the logical port names and
compares whether event with this logical port name occured and if so,
the waiting ends and port is removed from the cache.

This makes effective waiting for gateway router ports.

Change-Id: I5d2209906e70c0220ce20c4dc6593461d5dc7ffa
Closes-bug: #1842291