Temporarily switch master jobs to OVS 2.12 branch
While we adjust the jobs to build OVN from the new repo,
this patch is switching jobs that build OVS/OVN from master to
use the recently created branch-2.12.

In addition, make rally non-voting until
[1] is merged, the gate is broken there as well. So since there
are two separate issues breaking the gate, they can't depend on
each other to merge so fix it here. (It appears to be past
working hours for rally-core)

This patch must be reverted when [0] and [1] get merged.

[0] https://review.opendev.org/#/c/673236/
[1] https://review.opendev.org/#/c/681001/
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>

Change-Id: I63d12861e1ff5d444d116188a2a122eca5c45af0