Remove unneeded Zuul branch matcher
We have implicit branch matchers, so there's no need to add a check
for not-ocata etc, a job is only run for the branch it's on - like
master now.

Remove it to not confuse Zuul when multiple branches matches and the job
definition is different.

Remove jobs networking-ovn-tripleo-c7-s007-oooq-containers and
tripleo-ci-centos-7-scenario007-multinode-oooq-container, they are not run
on master at all, the jobs are on the branches and defined there.

Change-Id: If6e0b79ee4631ffa3295280c5ad04ad8a4399469