Test PortBindingChassisUpdateEvent
While other tests cover what happens when a Port_Binding is updated
when this event doesn't match, there was no test that covered the
case when it does. This test does test primarily implementation
details, but will have to do until we can test the actual use case
which involves uncleanly killing a compute server and doing a
'host evacuate' from it.

Change-Id: If9fc4209c5c54149a39ee3ddc743f3c11c51ceec