functional: Remove additional OVSDB connections
The functional test base created an explicit connections to the OVSDB
servers. However, there are other connections created by the mech_driver
itself. This patch replaces the explicit connections with the ones from
the mech_driver to avoid races in test between DB connections. e.g.
Connection1 creates resource and connection2 uses the resource - there
is a need every time to wait for connection2 to sync the DB to its local
replica.

Closes-bug: #1845003

Change-Id: Ib770a6c83192c06593463734d3f74f83eff513b4