Remove networking-ovn python2 jobs
Python 2 will be EOL at the end of 2019. Ussuri will no longer support
Python 2 so it's safe to remove python2 jobs and save some CI resources.

py27 unittest tox target is kept because the job is defined globally for all
projects. We can remove the py27 tox target once the job is removed.

Change-Id: I38f5a382bf54c30361ccc93d5f565b2cfdd1d621