[metadata agent] Account for 'external' ports
With the introduction of OVN 'external' ports, there can be
VIF ports that live in an external chassis. These should be
able to reach the metadata service so if an 'external' port
is bound to a hypervisor, metadata service will be provisioned
for it.

Co-Authored-By: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: Ib1c1eca68630f08de44aec29b6f6edac5e554f16
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit eb696ca703d1aaf168c54f37e0a25516e8953e5b)