Fix agent extension support after hashring merge
The agent API extension support made use of a cache that was filled
by processing events. Before hashring support, all events were
processed on each server, producing complete caches on those servers.
After distributing those events via the hashring, the cache for each
process would only have a partial list of the hash, so lookups for
agents would often fail. In addition, class on which the events were
registered was no longer being used, so the events wouldn't fire.

This patch deletes the in-memory AgentStats cache and instead stores
the values in the OVSDB and caches the chassis is memory based off of
the table contents instead of event triggers. This cache is required
in case ovn-controller is gracefully and temporarily shut down to keep
the Chassis from disappearing instead of showing up as "down".

Closes-Bug: #1853386
Closes-Bug: #1833394
Change-Id: I7a6f0f68a26381226686b410edb00dc6a56b5efb