tox: Pass OVN and OVS source directories
Some of networking-ovn functional tests inherit from ovsdbapp base
class. ovsdbapp uses environment variables to define directories
containing database schemas. This patch passes environment variable to
tox if it's defined. If it's not defined, an empty string is used and
code later uses its own defaults.

This patch doesn't go to master as it's frozen now because of the code
being merged to Neutron. The patch will be sent to master Neutron once
code migration is finished.

Change-Id: I5ccfc625352505c5b53f18e051e4d18598d5e286
Closes-bug: #1855903