Fix pep8 errors found by hacking 2.0.0
The following issues that are now exposed by hacking version 2.0.0

./networking_ovn/tests/unit/common/test_hash_ring_manager.py:92:38:
F602 dictionary key variable another_host_node repeated with different
values
 hash_dict_after_rebalance = {another_host_node: 'fake-uuid',

Also fixing

./networking_ovn/ovsdb/ovsdb_monitor.py:359:13: F841 local variable
'e' is assigned to but never used

which is already fixed in the neutron repo, so this is only an
issue in networking-ovn repo.

Also, this gerrit addresses an issue with functional tests where
ovsdbapp expects OVN schema to be located at '/usr/local/share/ovn/'.
So we are hard coding OVN_SRCDIR as /usr/local/share/openvswitch
to unblock the Zuul gate. The real fix is to make ovsdbapp smarter
about that directory, and then revert that change in tox.ini. This
particular issue is related to change
https://review.opendev.org/#/c/698293/ . Since functional tests are
not run for changes in tox.ini [1] so that fell through thecracks.

[1]: https://github.com/openstack/networking-ovn/blob/0d603a7e105d2cbe8ca21a1ee752e1f3c81879e0/zuul.d/legacy-networking-ovn-jobs.yaml#L10

Depends-On: https://review.opendev.org/699083/
Change-Id: I9396d8af17d281eef15777e70a7d8fc0791f20b2
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
Co-authored-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Co-authored-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Co-authored-by: Terry Wilson <twilson@redhat.com>
(cherry picked from commit b869f30e0b082d789c6668ec1a3ee69f2142798e)