Cache OvnProviderHelper object in OvnProviderDriver
Every time when a request comes from Octavia API and
is passed down to provider driver Octavia creates new
instance of Provider driver. This is good approach
while an API layer don't want to wait for previous API
actions to be end first, if centralized driver approach
will be used.

On different angle we have OvnProviderDriver and its
architecutre that was not created to work in dynamic
threads environment, but has centralized design.
Actual architecture looks like:

OvnProviderDriver object starts OvnPrividerHelper object
that manipulates OVN Databases. In OvnProviderHelper there is
a Queue.Queue object that stores actions to be executed
on databases with proper order. In addition OvnProviderHelper
starts dedicated thread that picks up actions from Queue
and execute those on OVN NorthBound database.

That solution does not scale while using it with Octavia
approach. There is a problem with thread leaking, because
in OvnProviderHelper we don't stop runner thread and we thread
it as daemon.

In addition on each API call the OVN Driver and dependent
components objects are created, this is big overhead.

To mitigate those architectural issues there is a possibility
to cache created OVNProviderHelper object and use one
instance of it for one Octavia API worker process.

This change must be cherry-picked to stable branches.

As next steps OVNProvider driver should be decomposed to:
   * Provider Agent as described in [1]
   * Provider Driver without Queue mechanism

[1] https://docs.openstack.org/octavia/latest/contributor/guides/providers.html#provider-agents

Change-Id: Ifaed77bbe6c36a63c18e568ec1d6ec5450dc8e03
Closes-Bug: 1856550