Don't fail in case subnet or Logical Switch not found
In case of subnet not found all the related actions could
not be taken. This patch prevent this situation to happen.
This could be possible that LB is going to be deleted after
the subnet (for example remove of the LoadBalancer in Error
state).

The same could happend with Logical Switch. Sometimes the
member network is delated in meantime and then we tries to
the members. It means we would need to remove LB to LS and
LB to LR associations, but it could not happend because of
missing LS. In those terms now we just pass the flow
and remove the LS to LR counters in external_ids.

Change-Id: I44f6de4c9a2b1c6054417dbb7f647a6af8339660
Closes-Bug: 1857009