OVN Metadata agent gets OVSDB updates for only its Chassis
The metadata agent registers the Chassis table with ovsdb-server
and therefor gets database updates every time *any* Chassis is
updated--even if the update is just a liveness check that updates
nb_cfg.

This patch adds a condition so that metadata agent only gets updates
for the Chassis that it is running on.

Change-Id: Ibce080d3af7d1c1c9b146ce5c4d9e96719e675df
Related-bug: #1861092
(cherry-picked from neutron I452b7de09312ecea621c4b448cc63f037cad9675)