Fix cirros image download during resources creation for migration
The link used for downloading cirros image is now redirecting to
some other location. Due to this reason the image was not downloaded
properly and created vm instance was not usable causing migration
to fail. This patch sets a correct link for cirros image. It also
adds a -L flag to curl allowing it to follow redirects and download
the image successfully.

Change-Id: Iebc8278aca56f87fcf74cc2725a1ffc1197c1bf4