Refactor OVN client QoS extension
The QoS OVN client extension is moved to the ML2 driver. This
extension is called from the OVN driver in the events of:
- create port
- update port
- delete port
- update network

The "update_policy" method is called from the OVN QoS driver, when
a QoS policy or its rules are updated.

The QoS OVN client extension updates the QoS OVN registers
exclusively, based on the related events.

NOTE: this patch is an non-straightforward cherry-pick from Neutron
(master branch) patch: https://review.opendev.org/#/c/711317/

Closes-Bug: #1863852

Change-Id: I7b7dca0318dfbe3690b58eccec593e84d08a98aa