Use rally-openstack 1.7.0 for stable/rocky
Master branch now has requirements incompatible with stable branches
constraints [0], so this switches to last tag compatible.

Longer term fix will be to run rally in a virtualenv, but this is lower
priority than other tasks. This can be revisited at the time, though for
older branches we will probably stick with pinned version

Also pin neutron-tempest-plugin to latest version that supports python 2

[0] https://opendev.org/openstack/rally-openstack/commit/5776e015f16321fdb489ebb482d984ccd3cd2903

Change-Id: I54fb021dec49724f17e7b5a330e09bcc638310d6
Closes-Bug: #1868691