Enable neutron-tempest-plugin installation in tempest venv
The neutron-tempest-plugin shouldn't be installed in global-scope
so thats why the tests weren't installed on stable gates.
We should enable the neutron-tempest-plugin as tempest-plugin, so that
the plugin will be installed in the py3 tempest venv.

In addition we can remove Octavia installation from base job
definition, because its not needed as we run the OVN Octavia
provider in separate job.

Added blacklist related to bug: 1881558.

Closes-Bug: #1881283
Related-Bug: #1881558

Change-Id: Ib7b92db86b77bc68fbec802d3642e9fa1ed5a57e