[OVN] HA Chassis Group: Ignore UPDATES when external_ids hasn't changed
The "old" parameter passed to the handle_ha_chassis_group_changes()
method is a delta object and sometimes it does not contain the
"external_ids" column (because it hasn't changed).

The absence of that column was misleading that method into believe that
the "old" object was no longer a gateway chassis and that triggered some
changes in the HA group. Changing the HA group resulted in the SRIOV
(external in OVN) ports to start flapping between the gateway chassis.

This patch is adding a check to verify that the "external_ids" column
has changed before acting on it, otherwise just ignore the update and
return.

Closes-Bug: #1869389
Change-Id: I3f7de633e5546dc78c3546b9c34ea81d0a0524d3
(cherry-picked manually from Neutron
 ea999564a5b80dcf13c0c43f107165f0754210b7)
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>