Clean up some of the OVN agent API methods
Adds ControllerAgent and MetadataAgent classes to organize some of
the OVN Agent API code.

Note from this backport: To be able to backport this to stable/train,
we had to remove the Python 3 only features used in the original patch
and make it compatible with python 2.7.

Change-Id: I9224edab2d27184ed96be1cdea2bd865aadff065
(cherry picked from Neutron 6723f4a485e2c60e45f45732c0881c126de06608)