[OVN] OVN driver to adapt to enable_distributed_floating_ip changes
Prior to this patch, if the [ovn]/enable_distributed_floating_ip
configuration option changed in an existing environment the OVN
driver wouldn't adapt to it requiring administrators to clear up the
"external_mac" column from the NAT table manually for the existing
floating ips.

With this patch, OVN will automatically correct existing NAT entries for
floating ips whenever this option changes.

To make the code simpler when handling the port up/down event this patch
always set the logical_port and the neutron:fip_external_mac key in the
external_ids column of the NAT entry when creating the floating ip.

Note that we are not using the maintenance task for this either, we are
re-using the event that set/unset the "external_mac" column for this
because, whenenver the service is restarted (after the configuration is
changed, we need to restart for it to take effect) the IDL will
re-trigger those events.

NOTE ABOUT THIS BACKPORT: Different from the master branch, this patch
also sets the "enable_distributed_floating_ip" in the
test_ovn_db_sync.py module. In the master branch, this was already set
prior to this patch.

(cherry-picked from Neutron I6a85fdde2558d781bf2853c5d11c5c964bbab81f)

Change-Id: Iebf658f44275fec948e34be439cf86092d2a782b
Closes-Bug: #1883559
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>