Fix the check for allowed_cidrs in listeners
The allowed_cidrs value could be an empty list if the
request involves the sdk, so change the check to
account for that.

Also had to set not_implemented_is_error to False
so we correctly skip tests with options the driver
does not support.

NOTE: This code is being added due to the recent addition
of an octavia-tempest-plugin test that could cause a
failure because the provider driver doesn't support this
option yet.

Related-bug: #1896603
(cherry picked from ovn-octavia-provider commit d43719bf3ccf4292740b140d33a5365bd2ce3fbf)

Change-Id: Ie1379747a62c5a656bf98f01c7814e73e506761f