Retry status updates to Octavia
When the OVN Octavia Provider driver calls into
octavia-lib to update the status of a loadbalancer,
it might fail. This can happen if a resource associated
with the loadbalancer was not found, for example, if
it's DB transaction was in-flight.

A side-effect is the resource can become "stuck":

$ openstack loadbalancer listener delete $ID
Load Balancer $ID is immutable and cannot be updated.

The provider driver needs to retry the operation,
typically even the very next call succeeds.

Conflicts:
    ovn_octavia_provider/helper.py

Change-Id: Ia4460dd0ae22949920f3f304b315e3c3f32f89c0
Closes-bug: #1900763
(cherry picked from ovn-octavia-provider commit 723ca5afeea3fa17e968d9c37220a4ff5a977d45)