Ensure ovsdb_probe_interval set before connect()
Setting the ovsdb_probe_interval after Connection.start() is
called means that the probe interval is not changed from
python-ovs's default of 5s until after the initial copy of the
database is retrieved. On busy systems, this can time out and
cause infinite reconnects.

This patch passes the probe_interval argument to the ovs.db.Idl
class so that it can be set as part of creating the jsonrpc
Session.

Some unit tests were removed and replaced with a functional test
which ensures not just that set_probe_interval is called, but that
the value is actually set before the connection is established.

Closes-bug: #1905611
(based on neutron commit 5783e952887b2fe5e8aaa32b3077e8dcdefa7252)

Change-Id: I54ede5de8cbfc5ac57a536dd1113be1213edc6e4