[OVN] Ensure metadata checksum
Ensure the TCP traffic leaving the OVN metadata namespace has
the checksum correctly populated. This is necessary when the
OVS datapath is "netdev".

Because the overhead added is minimal and only applies to the
metadata traffic inside the metadata namespace, this rule is
always set.

NOTE: this patch also includes [1], that fixes a bug introduced
by the original patch.

[1]https://review.opendev.org/c/openstack/neutron/+/772286

Manually cherry-picked from
https://review.opendev.org/c/openstack/neutron/+/763745/12

Closes-Bug: #1904871

Change-Id: Ib6eb02100cedc152ce589e67c0e7da74b26a6218