Catch subnet not found exceptions better
The Octavia Neutron driver code might not return the
neutronclient NotFound exception cleanly, so make the
code broader and look for the "subnet not found" string.

Moved to a common function to not duplicate code.

NOTE: This is not required in the OVN Octavia Provider
Driver code for Ussuri and newer as it uses the neutron
client directly.

Change-Id: I7e34db30a7ffb257c7574e75a64a886ae4e5aca5