[OVN] Set mcast_flood_reports on LSPs
This patch enables the "mcast_flood_reports" and "mcast_flood" (on
provnet ports only) options in the Logical Switch Ports in the OVN driver.
Without these options, the ovn-controller will consume the IGMP queries
and won't send it to the LSP ports, meaning that external IGMP queries
will never arrive to the VMs.

In talks to the core OVN team, it was suggested [0] to enable the
"mcast_flood_reports" option by default in the OVN driver (at least until
fixed in core OVN) as a workaround to this problem. And, to avoid having
to update all ports (which can be many) based on the igmp_snooping_enable
configuration option, we are always setting "mcast_flood_reports" to
"true" in the LSPs. This won't cause any harm (also confirmed by core OVN
developers [0]) since it will be ignored if multicast snoop is disabled.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1933990#c3

(Manually cherry picked from Neutron b04d64b90f2822222f96ddd726d379a2f19b539e)

Closes-Bug: #1918108
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: I63853f1b01c9b7444ed592c871a60125fa64f098