[OVN] Set send_periodic to False on provider networks
This patch adds a check prior to setting the send_periodic configuration
option for the LRP's ipv6_ra_configs to see if the network the LRP is
connected to is a provider network and the port is a gateway port.

In case it is a provider network and the port is a gateway port,
send_periodic should be set to False to avoid leaking the RAs generated
by ovn-controller to the tenant networks onto the provider network.

(Manually cherry picked from Neutron 1f68336b79b0432308cdee42ea39ede49707dfa3)

Closes-Bug: #1919347
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: I9781a32483841d8e63062514f6e3641870d4d4f7