[ovn] Add neutron network to metadata namespace names
Until this patch the metadata namespace had the format
'ovnmeta-<OVN datapath uuid>' which makes it hard to
relate to the Neutron network for which metadata is
provisioned to. This applies to the name of the tap
interface that connects the network namespace to the
integration bridge.

This patch is changing the name convention and providing
an upgrade path to include the Neutron network ID to make
debugging and troubleshooting easier for developers and
operators.

The new name pattern is:
'ovnmeta-<Neutron network uuid>'

Please, note that with this patch, the old namespaces will
be deleted and new ones will be recreated.

(Manually cherry picked from Neutron e4fb06b24299a7ecf10b05ef6ddc2d883c40e5a1)

Conflicts:
    networking_ovn/agent/metadata/agent.py
    networking_ovn/common/utils.py
    networking_ovn/ovsdb/impl_idl_ovn.py

Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
Change-Id: Ic8ffa9c4437aab6fb1878b3a1ebf2c3ab86e3d0c