Mark scenario trunk tests as unstable for now for python 2 job
It's failing from time to time and it's reported in the related bug.
So lets mark this test as unstable for now, until we will fix the issue
to make our gate more stable.

As the parent job (networking-ovn-base) already had a tempest_black_regex,
set, this child job explicitly adds the parent's regex to the list. Which
means the IPv6Test is also part of
networking-ovn-tempest-dsvm-ovs-release-python2's tempest_black_regex.

Related-Bug: #1927985

Change-Id: I49df20960fbb142ca5dd1affe715d2a7bd580bcf