Use TCP keepalives for ovsdb connections
When failing over OVN DB servers from one server to another, the
server which originally hosted the VIP doesn't notice the connection
is gone and doesn't reconnect. Ultimately, this is something that
needs to be fixed in python-ovs, but setting the SO_KEEPALIVE socket
option avoids the issue. This also has the benefit that the client
doesn't need to send 'echo' requests, which can time out on an
overloaded ovsdb-server, which causes a disconnection which then
adds even more load on the ovsdb-server as it has to send the entire
db contents over the wire after the connection.

Closes-Bug: #1930926
Change-Id: Ie0205785cab307c132fbe409588739685cade7c0
(cherry picked from neutron commit 65cce351d74a9a637fdb2a9d5e0e63445dda9ea9)