Fix neutron_pg_drop-related startup issues
There is a python-ovs bug that allows transactions to be executed
before it has started monitoring the db for changes. This breaks
ovsdbapp's AddCommand behavior, which looks up a row in memory
post-commit to return it to the caller, because we process the
reply from ovsdb-server that has the UUID in it, but when we try to
look it up in memory, it isn't there since that is handled via the
monitor notifications.

Since we don't care about the return value, we can just ignore the
KeyError. This is a slightly different fix than the one in neutron
but it does the same thing.

Closes-Bug: #1938766
Change-Id: I78af15970a86923fb810e903d38634b6e60a8fe4